+++ packages/wallet/src/redux/protocols/challenging/challenger/components/approve-challenge.tsx
