+++ packages/wallet/src/redux/protocols/challenging/challenger/index.ts
@@ -1,0 +1,2 @@
+export { Challenger as Challenging } from './container';
+export { initialize, challengerReducer as reducer } from './reducer';
