--- packages/wallet/src/redux/protocols/challenging/index.ts
@@ -1,2 +1,0 @@
-export { Challenging } from './container';
-export { initialize, challengingReducer as reducer } from './reducer';
