+++ packages/wallet/src/redux/protocols/challenging/responder/components/wait-for-approval.tsx
