+++ packages/wallet/src/redux/protocols/challenging/responder/index.ts
@@ -1,0 +1,1 @@
+export { initialize, responderReducer as reducer } from './reducer';
