+++ packages/wallet/src/redux/protocols/challenging/state.ts
@@ -1,0 +1,36 @@
+import {
+  ResponderState,
+  TerminalResponderState,
+  isTerminal as isResponderTerminal,
+  isResponderState,
+  NonTerminalResponderState,
+} from './responder/state';
+import {
+  ChallengerState,
+  TerminalChallengerState,
+  isTerminal as isChallengerTerminal,
+  isChallengerState,
+  NonTerminalChallengerState,
+} from './challenger/states';
+import { ProtocolState } from '..';
+
+export type ChallengingState = ResponderState | ChallengerState;
+
+export function isTerminal(
+  state: ChallengingState,
+): state is TerminalChallengerState | TerminalResponderState {
+  return (
+    (isChallengerState(state) && isChallengerTerminal(state)) ||
+    (isResponderState(state) && isResponderTerminal(state))
+  );
+}
+
+export function isChallengingState(state: ProtocolState): state is ChallengerState {
+  return isChallengerState(state) || isResponderState(state);
+}
+
+export function isNonTerminalChallengingState(
+  state: ProtocolState,
+): state is NonTerminalChallengerState | NonTerminalResponderState {
+  return isChallengingState(state) && !isTerminal(state);
+}
