+++ packages/wallet/src/redux/protocols/container.tsx
@@ -1,7 +1,7 @@
 import { PureComponent } from 'react';
 import { ProtocolState } from '.';
 import * as fundingStates from './funding/states';
-import * as challengingStates from './challenging/states';
+import * as challengingStates from './challenging/state';
 import * as concludingStates from './concluding/state';
 import React from 'react';
 import { Funding } from './funding/container';
@@ -9,9 +9,8 @@ import { Concluding } from './concluding/container';
 import { connect } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
-import { Challenging } from './challenging';
-import * as respondingStates from './responding/state';
-import { Responding } from './responding/container';
+import { Challenging } from './challenging/container';
+
 interface Props {
   protocolState: ProtocolState;
 }
@@ -24,10 +23,8 @@ class ProtocolContainer extends PureComponent<Props> {
     const { protocolState } = this.props;
     if (fundingStates.isNonTerminalFundingState(protocolState)) {
       return <Funding state={protocolState} />;
-    } else if (challengingStates.isNonTerminalChallengingState(protocolState)) {
+    } else if (!challengingStates.isNonTerminalChallengingState(protocolState)) {
       return <Challenging state={protocolState} />;
-    } else if (respondingStates.isNonTerminalRespondingState(protocolState)) {
-      return <Responding state={protocolState} />;
     } else if (concludingStates.isConcludingState(protocolState)) {
       return <Concluding state={protocolState} />;
     } else {
