--- packages/wallet/src/redux/protocols/responding/index.ts
@@ -1,1 +1,0 @@
-export { initialize, respondingReducer as reducer } from './reducer';
